Pulls all string-related methods into another module to satisfy codeclimate.
